Tidy-up, Timing file & minor changes
Added more-proper comments and intoduction of a timing header file for readout function. This includes comipler readout control in BoidSimOMP.cpp.

+++ cpp_omp/Boid.h
@@ -23,4 +23,4 @@ void UpdatePos(Boid&);
 void Reset(Boid&, int);
 
 // Flock functions
-std::vector<Boid> GenFlock(int, double, double, double, double);
+std::vector<Boid> GenFlock(int, int, int, int, int);
